small fixes - the server runs now

+++ src/Simulator/DelayedFluidSimulator.h
@@ -59,8 +59,8 @@ protected:
 	virtual cFluidSimulatorData * CreateChunkData(void) override { return new cDelayedFluidSimulatorChunkData(m_TickDelay); }
 
 	int m_TickDelay;   // Count of the m_Slots array in each ChunkData
-	int m_AddSlotNum;  // Index into m_Slots[] where to add new blocks in each ChunkData
-	int m_SimSlotNum;  // Index into m_Slots[] where to simulate blocks in each ChunkData
+	size_t m_AddSlotNum;  // Index into m_Slots[] where to add new blocks in each ChunkData
+	size_t m_SimSlotNum;  // Index into m_Slots[] where to simulate blocks in each ChunkData
 
 	int m_TotalBlocks;  // Statistics only: the total number of blocks currently queued
